requirements.txt
streamlit
Pillow
tensorflow
tensorflow_hub
numpy
